Uploaded start of pigeon shooter mini game
Added three new scenes: one for the main screen of the mini game, one for clothing that moves in a projectile arch and finally one for projectile. The player can move left and right, shoot poop with spacebar (this means I've added actions in project settings as well), clothing spawns after a time interval that gets smaller and smaller and a countdown starts the game.