Expanded pigeon shooter
Added stat dependancy for CON and CHA. There is now egg chance, dps depending on CON, a cool down meter to show when poop is ready. The arch is also fixed, and clothing now lands on the clothing line. I've also added an ending animation with text and a return button to the hotel scene.